@@ -106,8 +106,8 @@
     }
   },
   "devDependencies": {
-    "@babel/core": "^7.15.5",
-    "@babel/preset-react": "^7.14.5",
+    "@babel/core": "^7.17.8",
+    "@babel/preset-react": "^7.16.7",
     "@electron-forge/cli": "^6.0.0-beta.60",
     "@electron-forge/maker-deb": "^6.0.0-beta.60",
     "@electron-forge/maker-dmg": "*",
@@ -118,11 +118,11 @@
     "@electron-forge/publisher-github": "*",
     "@pmmmwh/react-refresh-webpack-plugin": "^0.5.4",
     "@vercel/webpack-asset-relocator-loader": "^1.7.0",
-    "babel-loader": "^8.2.2",
+    "babel-loader": "^8.2.4",
     "concurrently": "^6.5.1",
     "cross-env": "^7.0.3",
     "css-loader": "^6.2.0",
-    "electron": "^17.0.0",
+    "electron": "^18.0.1",
     "electron-installer-common": "^0.10.3",
     "eslint": "^7.32.0",
     "eslint-config-airbnb": "^18.2.1",
@@ -135,15 +135,15 @@
   },
   "dependencies": {
     "@chakra-ui/icons": "^1.0.15",
-    "@chakra-ui/react": "^1.8.6",
-    "@emotion/react": "^11.8.1",
+    "@chakra-ui/react": "^1.8.7",
+    "@emotion/react": "^11.8.2",
     "@emotion/styled": "^11.8.1",
     "@fontsource/open-sans": "^4.5.1",
     "@geoffcox/react-splitter": "^2.1.1",
     "@react-hook/window-size": "^3.0.7",
     "@react-nano/use-event-source": "^0.11.0",
     "decompress": "^4.2.1",
-    "electron-log": "^4.4.1",
+    "electron-log": "^4.4.6",
     "electron-squirrel-startup": "^1.0.0",
     "framer-motion": "^6.2.8",
     "html-to-image": "^1.9.0",
@@ -152,11 +152,11 @@
     "os-utils": "^0.0.14",
     "portfinder": "^1.0.28",
     "re-resizable": "^6.9.5",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
+    "react": "^18.0.0",
+    "react-dom": "^18.0.0",
     "react-flow-renderer": "^10.0.7",
     "react-hotkeys-hook": "^3.4.4",
-    "react-icons": "^4.2.0",
+    "react-icons": "^4.3.1",
     "semver": "^7.3.5",
     "systeminformation": "^5.9.15",
     "use-debounce": "^7.0.1",
