@@ -1,7 +1,10 @@
 import React from 'react';
-import ReactDOM from 'react-dom';
+import * as ReactDOMClient from 'react-dom/client';
 // eslint-disable-next-line import/extensions
 import App from './app.jsx';
 import './global.css';
 
-ReactDOM.render(<App />, document.getElementById('root'));
+const container = document.getElementById('root');
+const root = ReactDOMClient.createRoot(container);
+
+root.render(<App />);
